Created Navigation and Altered File Structure
Updated dependencies: added react-router and react-router-dom to project.

Updated index.js: added the dark theme to the index file to give the project a new look.

Added filler info for files in order to upload the file structure: Admin, Market, Portfolio, and UserTransaction files.

Updated App.js: removed initial data in file and replaced with Nav and Router. Used components to access other files and display information from this file. Added routes to properly connect the project to the right files when the nav is accessed.

Small changes to index.js:  added bootstrap import to the proper file and linked the correct component.

Removed testy.js as it is no longer needed.